requirements.txt
flask
agno
python-dotenv